new animation  & fixed splash layout
